+++ docs/source/changelog.rst
@@ -7,6 +7,7 @@ Version 0.6.18 (2020-09-xx)
 - Reduced memory consumption when building gzip index
 - Improved the speed of random access to reads from fastq
 - Fixed sequence dealloc error cuasing no fasta delloc trigger
+- Fixed fastq max and min quality score return value
 
 Version 0.6.17 (2020-08-31)
 ---------------------------